refactor: rename filtersUrl to filtering, rm extra test file
