+++ src/components/CurriculumComponents/CurricVisualiserFilters/CurricFiltersThreads.tsx
@@ -5,7 +5,7 @@ import { Fieldset, FieldsetLegend } from "../OakComponentsKitchen/Fieldset";
 import { RadioGroup, RadioButton } from "../OakComponentsKitchen/SimpleRadio";
 
 import { Thread, CurriculumFilters } from "@/utils/curriculum/types";
-import { highlightedUnitCount } from "@/utils/curriculum/filtersUrl";
+import { highlightedUnitCount } from "@/utils/curriculum/filtering";
 import { CurriculumUnitsFormattedData } from "@/pages-helpers/curriculum/docx/tab-helpers";
 import { joinWords, pluralizeUnits } from "@/utils/curriculum/formatting";
 import { ComponentTypeValueType } from "@/browser-lib/avo/Avo";
